SSG: high volume performance (#668)
## The problem

HTML payloads of static pages grow linearly with the number of them.
This is already evident in the `03_demo` example. Each pokemon has an
entry in `__WAKU_ROUTER_PREFETCH__` that points to the identical
component. The `/SHOULD_SKIP` entry also grows. With a larger number of
pages, the build breaks due to memory issues or node simply not being
able to `JSON.stringify` the build config in:

https://github.com/dai-shi/waku/blob/6550a0d26e7830527c767fcda8eb906aeb0f1266/packages/waku/src/lib/builder/build.ts#L692

Also, when generating the map for `__WAKU_ROUTER_PREFETCH__`, pages are
rendered non-concurrently, which means blocking operations can't run in
parallel.

## Solutions

1. I switched router prefetching to a pattern instead of path-based
model. That way each static page needs only one entry. Not one for each
static path.
2. The `shouldSkipObj` grows with each page entry. To be completely
honest, I don't know enough about the rendering process to judge if each
page would need skip lists for each other page component. But that
growth during rendering felt weird, so i just moved the object one scope
level down to reset it for each page. It's a complete shot in the dark,
but it did hit something.

https://github.com/dai-shi/waku/blob/6550a0d26e7830527c767fcda8eb906aeb0f1266/packages/waku/src/router/define-router.ts#L96
3. Collection of modules for `__WAKU_ROUTER_PREFETCH__` has been changed
to run in parallel. This is also necessary to enable in-component
optimizations, like database request batching or connection pooling.

## Notes

* The `ssg-performance` fixture renders 10k pages in about a minute on
my M1. Before it would just break.
* The build config serialization might still break at some threshold,
due to `JSON.stringify` limits.
* Rendering all static paths to collect modules for prefetching feels
wasteful, but thats too big of a change for this PR. But this brings us
into a working state.

---------

Co-authored-by: daishi <daishi@axlight.com>
Co-authored-by: Daishi Kato <dai-shi@users.noreply.github.com>