fix: Hono runner 404 response (#796)
This fixes the issue for me mentioned in #644 where an empty 200 status
response is returned instead of a 404 error page.

I found the RSC middleware doesn't run in that case so the ctx.res is
never set. I added a fallback to move the hono res.body to the ctx.body
if the ctx.body is empty. I assume that won't be a breaking change but
🤷.

---------

Co-authored-by: daishi <daishi@axlight.com>