feat: cloudflare pages adapter, alternate implementation (#795)
Here is an alternate attempt at Cloudflare Pages support. Related to
#793.

- In output-cloudflare.ts, it moves the outDir to a subdir named
`worker.js` (a directory, not a file). It creates a
dist/worker.js/index.js file to load the `dist/worker.js/serve.js` file.
- It creates a _routes.json file in the root if one doesn't already
exist from the publicDir and writes the publicDir entries to it.
- It moves the publicDir to the root of the outDir.
- It outputs an updated wrangler.toml if one doesn't exist with basic
settings
- It uses env.ASSETS.fetch to load a custom 404.html file.

Reference:

-
https://developers.cloudflare.com/pages/functions/wrangler-configuration/
-
https://developers.cloudflare.com/pages/functions/api-reference/#envassetsfetch
-
https://github.com/withastro/adapters/blob/main/packages/cloudflare/src/index.ts
-
https://github.com/sveltejs/kit/blob/main/packages/adapter-cloudflare/index.js
-
https://github.com/honojs/vite-plugins/blob/main/packages/cloudflare-pages/src/cloudflare-pages.ts

A waku plugin can be used to make wrangler dev server bindings
available.

-
https://developers.cloudflare.com/workers/wrangler/api/#getplatformproxy

This _could_ be bundled into waku or an external plugin... but it would
want the developer to supply the wrangler version as a peer dependency.
Cloudflare is constantly publishing wrangler and miniflare versions with
new features. I published an example in a gist:
https://gist.github.com/rmarscher/9bb6ed54dc9535f4b81bed147204c7e9

I published examples/10_fs-router to
https://c03fa4b1.waku-test.pages.dev/

Thanks for waku! I'm looking forward to trying it on my cloudflare pages
projects!

---------

Co-authored-by: Daishi Kato <dai-shi@users.noreply.github.com>