feat: expose hash from `useRouter` (#746)
## What was changed?

Expose the current url hash via `UNSTABLE_useRouter`. On the server, its
always an empty string, but gets populated on the client after the
initial render.

## Why was it changed?

Otherwise inconvenient `useEffect`-gymnastics are necessary to get the
current hash reliably into a client component.

## Notes

In the process of writing tests I found another potential issue with
`searchParams`.


https://github.com/dai-shi/waku/pull/746/files#diff-a8dab8847d02ba73c6b66e738495556ddda5d2dd01b70322aae006f7669ea237R6-R11

I'm not sure if we should move the double-creation of `URLSearchParams`
into the `useRouter` hook. Given some advice, I can try to fix that as
well.

---------

Co-authored-by: Daishi Kato <dai-shi@users.noreply.github.com>
Co-authored-by: daishi <daishi@axlight.com>