fix: cloudflare routing 404 (#870)
This fixes cloudflare static vs dynamic routing for the home index page.
But there are likely more bugs with the _routes.json when trying to nest
a dynamic route under a static route path. By using _routes.json, we can
route static html files directly off Cloudflare's CDN and not incur any
worker function changes. But we need to make sure it only matches the
static html routes.

Right now, it is doing a very basic walk of the public output dir to
generate a _routes.json. We need to generate a more accurate list of
exclude patterns. I am opening this PR as a draft until I can complete
this. Fixes #869