fix(deploy): move saving the build config to deploy plugins (#888)
This (along with the previous #874) resolves the issue in #871. The
build config was being written to the wrong location for cloudflare
pages This PR exports a function from the build module that appends the
build config to the given entries file location. It uses a new unstable
build config state variable (`unstable_buildConfigSave`) to track
whether or not the build config has be saved to the entries file. This
allows the cloudflare deploy plugin to write the build config into the
worker function directory instead of the root.

We could merge #879 and then I can edit the base branch to main and
continue with this PR. Or however you want to handle it. Thank you!