idea: new defineRouter (#949)
We would like to redesign `defineRouter`. This is why we have
`unstable_` prefix.

https://github.com/dai-shi/waku/issues/885#issuecomment-2395208298

> similar snippets in both `defineRouter` and `createPages` are
currently assuming 1:1 relationship from path to component id.

That's the exactly one of the reasons. It shouldn't be 1:1. It can be
anything, and it can cache each. This will change waku/router/client too
including "shouldSkipList".

Starting with a rough idea.