fix(style-dev): style css for dev with vite js module
This will fix the missing styles with vite 6 for #1012

It works by running js from the browser to create the styles with CSR
in dev. Before we were extracting the source for plain css and passing those
from the dev server in a <style> tag, but since that breaks in vite 6,
this is an easy solution that will handle styles more consistently and without
errors.