fix: netlify relative fs path
Changes the netlify functions output dir from netlify/functions
to netlify-functions.

This is a breaking change for existing netlify projects
using dynamic rendering. It requires updating netlify.toml `[functions]` section with

```
  directory = "netlify-functions"
```

In a monorepo, this needs to also include the base path from the root of the monorepo.

Fixes #909

to the netlify.toml (or to