fix(router): decode path for storing elements and slug as prop
Let me know if this approach is ok. The change makes it so that id's
internally for routes are always decoded. This makes it so the prop
is decoded when it is passed. It also fixes the runtime error. The runtime
error happened when the id as being looked up in elements as decoded, but
internally stored as encoded, so our invalid element id error was thrown.

fixes #1302